requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
pandas==2.1.1
numpy==1.27.5
scikit-learn==1.3.2
xgboost==1.7.7
sqlalchemy==2.0.20
pymysql==1.1.1
python-dotenv==1.0.0
joblib==1.3.2
